+++ xcube/util/cmaps.py
@@ -1,283 +1,421 @@
 # The MIT License (MIT)
-# Copyright (c) 2019 by the xcube development team and contributors
+# Copyright (c) 2022 by the xcube team and contributors
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy of
-# this software and associated documentation files (the "Software"), to deal in
-# the Software without restriction, including without limitation the rights to
-# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
-# of the Software, and to permit persons to whom the Software is furnished to do
-# so, subject to the following conditions:
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
 #
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
 #
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 # FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 # AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+# DEALINGS IN THE SOFTWARE.
+
 
 import base64
 import io
 import os
 import re
 from functools import cached_property
-from typing import Dict, Tuple, List, Optional, Set
+from typing import Dict, Tuple, List, Optional
 
 import matplotlib
-import matplotlib.cm as cm
 import matplotlib.colors
 import numpy as np
 from PIL import Image
+from deprecated import deprecated
 
+from util.assertions import assert_in, assert_instance
 from xcube.constants import LOG
 
 try:
-    import cmocean.cm as ocm
+    # noinspection PyPackageRequirements
+    import cmocean.cm as cmo
 except ImportError:
-    ocm = None
+    cmo = None
 
-__author__ = "Norman Fomferra (Brockmann Consult GmbH)"
+OCEAN_PREFIX = "cmo"
+ALPHA_SUFFIX = "_alpha"
+REVERSED_SUFFIX = "_r"  # Used by ocm / Ocean
+
+DEFAULT_CMAP_NAME = "viridis"
 
 
-# Have colormaps separated into categories:
-# (taken from http://matplotlib.org/examples/color/colormaps_reference.html)
-# colormaps for ocean:
-# (taken from https://matplotlib.org/cmocean/)
+# Have colormaps separated into categories taken from
+# https://matplotlib.org/stable/gallery/color/colormap_reference.html
+# colormaps for ocean taken from
+# https://matplotlib.org/cmocean/
 
 
 class ColormapCategory:
     def __init__(self,
                  name: str,
                  desc: str,
-                 cmap_names: Optional[Set[str]] = None):
+                 cm_names: Optional[List[str]] = None):
         self.name = name
         self.desc = desc
-        self.cmap_names = cmap_names
+        self.cm_names = cm_names
 
 
-PERCEPTUALLY_UNIFORM_SEQUENTIAL = ColormapCategory(
+PERCEPTUALLY_UNIFORM_SEQUENTIAL_CATEGORY = ColormapCategory(
     'Perceptually Uniform Sequential',
     'For many applications, a perceptually uniform colormap'
     ' is the best choice - one in which equal steps in data are'
     ' perceived as equal steps in the color space',
-    {'viridis', 'inferno', 'plasma', 'magma'}
+    [
+        'viridis', 'plasma', 'inferno', 'magma', 'cividis'
+    ]
 )
 
-SEQUENTIAL_1 = ColormapCategory(
-    'Sequential 1',
+SEQUENTIAL_CATEGORY = ColormapCategory(
+    'Sequential',
     'These colormaps are approximately monochromatic colormaps varying'
     ' smoothly between two color tones - usually from low saturation'
     ' (e.g. white) to high saturation (e.g. a bright blue).'
     ' Sequential colormaps are ideal for representing most scientific'
     ' data since they show a clear progression from'
     ' low-to-high values.',
-    {'Blues', 'BuGn', 'BuPu',
-     'GnBu', 'Greens', 'Greys', 'Oranges', 'OrRd',
-     'PuBu', 'PuBuGn', 'PuRd', 'Purples', 'RdPu',
-     'Reds', 'YlGn', 'YlGnBu', 'YlOrBr', 'YlOrRd'}
+    [
+        'Greys', 'Purples', 'Blues', 'Greens', 'Oranges', 'Reds',
+        'YlOrBr', 'YlOrRd', 'OrRd', 'PuRd', 'RdPu', 'BuPu',
+        'GnBu', 'PuBu', 'YlGnBu', 'PuBuGn', 'BuGn', 'YlGn'
+    ]
 )
 
-SEQUENTIAL_2 = ColormapCategory(
-    'Sequential 2',
+SEQUENTIAL_2_CATEGORY = ColormapCategory(
+    'Sequential (2)',
     'Many of the values from the Sequential 2 plots are monotonically'
     ' increasing.',
-    {'afmhot', 'autumn', 'bone', 'cool',
-     'copper', 'gist_heat', 'hot',
-     'pink', 'spring', 'summer', 'winter'}
+    [
+        'binary', 'gist_yarg', 'gist_gray', 'gray', 'bone', 'pink',
+        'spring', 'summer', 'autumn', 'winter', 'cool', 'Wistia',
+        'hot', 'afmhot', 'gist_heat', 'copper'
+    ]
 )
 
-DIVERGING = ColormapCategory(
+DIVERGING_CATEGORY = ColormapCategory(
     'Diverging',
     'These colormaps have a median value (usually light in color) and vary'
     ' smoothly to two different color tones at high and low values.'
     ' Diverging colormaps are ideal when your data has a median value'
     ' that is significant (e.g.  0, such that positive and negative'
     ' values are represented by different colors of the colormap).',
-    {'BrBG', 'bwr', 'coolwarm', 'PiYG', 'PRGn', 'PuOr',
-     'RdBu', 'RdGy', 'RdYlBu', 'RdYlGn', 'Spectral',
-     'seismic'}
+    [
+        'PiYG', 'PRGn', 'BrBG', 'PuOr', 'RdGy', 'RdBu',
+        'RdYlBu', 'RdYlGn', 'Spectral', 'coolwarm', 'bwr', 'seismic'
+    ]
 )
-QUALITATIVE = ColormapCategory(
+
+QUALITATIVE_CATEGORY = ColormapCategory(
     'Qualitative',
     'These colormaps vary rapidly in color.'
     ' Qualitative colormaps are useful for choosing'
     ' a set of discrete colors.',
-    {'Accent', 'Dark2', 'Paired', 'Pastel1',
-     'Pastel2', 'Set1', 'Set2', 'Set3'}
+    [
+        'Pastel1', 'Pastel2', 'Paired', 'Accent',
+        'Dark2', 'Set1', 'Set2', 'Set3',
+        'tab10', 'tab20', 'tab20b', 'tab20c'
+    ]
 )
 
-UNCATEGORIZED = ColormapCategory(
-    'Uncategorized',
+CYCLIC_CATEGORY = ColormapCategory(
+    'Cyclic',
+    'These colormaps are cyclic.',
+    [
+        'twilight', 'twilight_shifted', 'hsv'
+    ]
+)
+
+MISCELLANEOUS_CATEGORY = ColormapCategory(
+    'Miscellaneous',
     'Colormaps that don\'t fit into the categories above'
-    ' or haven\'t been categorized yet.'
+    ' or haven\'t been categorized yet.',
+    [
+        'flag', 'prism', 'ocean', 'gist_earth', 'terrain', 'gist_stern',
+        'gnuplot', 'gnuplot2', 'CMRmap', 'cubehelix', 'brg',
+        'gist_rainbow', 'rainbow', 'jet', 'turbo', 'nipy_spectral',
+        'gist_ncar'
+    ]
 )
 
-OCEAN = ColormapCategory(
+OCEAN_CATEGORY = ColormapCategory(
     'Ocean',
     'Colormaps for commonly-used oceanographic variables'
     ' (from module cmocean.cm).'
 )
 
-CUSTOM = ColormapCategory(
-    'Custom Colormaps',
+CUSTOM_CATEGORY = ColormapCategory(
+    'Custom',
     'Custom colormaps, e.g. loaded from SNAP *.cpd files.',
 )
 
-_CATEGORIES: Dict[str, ColormapCategory] = {
+TEMPLATE_CATEGORIES: Dict[str, ColormapCategory] = {
     c.name: c
     for c in (
-        PERCEPTUALLY_UNIFORM_SEQUENTIAL,
-        SEQUENTIAL_1,
-        SEQUENTIAL_2,
-        DIVERGING,
-        QUALITATIVE,
-        UNCATEGORIZED,
-        OCEAN,
-        CUSTOM,
+        PERCEPTUALLY_UNIFORM_SEQUENTIAL_CATEGORY,
+        SEQUENTIAL_CATEGORY,
+        SEQUENTIAL_2_CATEGORY,
+        DIVERGING_CATEGORY,
+        QUALITATIVE_CATEGORY,
+        CYCLIC_CATEGORY,
+        OCEAN_CATEGORY,
+        CUSTOM_CATEGORY,
+        MISCELLANEOUS_CATEGORY,
     )
 }
 
-_CMAP_NAME_TO_CAT = {
-    cm_name: cat
-    for cat in _CATEGORIES.values()
-    if cat.cmap_names
-    for cm_name in cat.cmap_names
+_CM_NAME_TO_TEMPLATE_CATEGORY = {
+    cm_name: template_cat
+    for template_cat in TEMPLATE_CATEGORIES.values()
+    if template_cat.cm_names
+    for cm_name in template_cat.cm_names
 }
 
 
 class Colormap:
     def __init__(self,
-                 cat_name: str,
-                 cmap_name: str,
-                 cmap: matplotlib.colors.Colormap,
+                 cm_name: str,
+                 cat_name: Optional[str] = None,
+                 cmap: Optional[matplotlib.colors.Colormap] = None,
+                 cmap_reversed: Optional[matplotlib.colors.Colormap] = None,
                  cmap_alpha: Optional[matplotlib.colors.Colormap] = None,
-                 cmap_reverse: Optional[matplotlib.colors.Colormap] = None):
+                 norm: Optional[matplotlib.colors.Normalize] = None):
+        self._cm_name = cm_name
         self._cat_name = cat_name
-        self._cmap_name = cmap_name
         self._cmap = cmap
+        self._cmap_reversed = cmap_reversed
         self._cmap_alpha = cmap_alpha
-        self._cmap_reverse = cmap_reverse
+        self._cmap_reversed_alpha = None
         self._cmap_png_base64: Optional[str] = None
+        self._norm = norm
 
     @property
-    def cat_name(self) -> str:
+    def cm_name(self) -> str:
+        return self._cm_name
+
+    @property
+    def cat_name(self) -> Optional[str]:
         return self._cat_name
 
-    @property
-    def cmap_name(self) -> str:
-        return self._cmap_name    \
-
-    @property
+    @cached_property
     def cmap(self) -> matplotlib.colors.Colormap:
+        if self._cmap is None:
+            self._cmap = matplotlib.colormaps[self.cm_name]
         return self._cmap
 
     @cached_property
     def cmap_alpha(self) -> matplotlib.colors.Colormap:
         if self._cmap_alpha is None:
-            self._cmap_alpha = _get_alpha_cmap(self.cmap)
+            _, self._cmap_alpha = get_alpha_cmap(
+                self.cm_name, self.cmap
+            )
         return self._cmap_alpha
 
     @cached_property
-    def cmap_reverse(self) -> matplotlib.colors.Colormap:
-        if self._cmap_reverse is None:
-            self._cmap_reverse = self.cmap.reversed(
-                name=self._cmap_name + "_r"
+    def cmap_reversed(self) -> matplotlib.colors.Colormap:
+        if self._cmap_reversed is None:
+            _, self._cmap_reversed = get_reverse_cmap(
+                self.cm_name, self.cmap
             )
-        return self._cmap_reverse
+        return self._cmap_reversed
+
+    @cached_property
+    def cmap_reversed_alpha(self) -> matplotlib.colors.Colormap:
+        if self._cmap_reversed_alpha is None:
+            cm_name, cmap_reversed = get_reverse_cmap(
+                self.cm_name, self.cmap
+            )
+            _, self._cmap_reversed_alpha = get_alpha_cmap(
+                cm_name, cmap_reversed
+            )
+        return self._cmap_reversed_alpha
 
     @cached_property
     def cmap_png_base64(self) -> str:
         if self._cmap_png_base64 is None:
-            self._cmap_png_base64 = _get_cmap_png_base64(self.cmap)
+            self._cmap_png_base64 = get_cmap_png_base64(self.cmap)
         return self._cmap_png_base64
 
+    @property
+    def norm(self) -> Optional[matplotlib.colors.Normalize]:
+        return self._norm
 
-# def get_cmap(cmap_name: str,
-#              default_cmap_name: str = 'viridis',
-#              num_colors: Optional[int] = None) -> CategorizedColormaps:
 
-class CategorizedColormaps:
-    def __init__(self):
-        pass
-
-    def _load(self):
-        categories: Dict[str, ColormapCategory] = {}
-        cmaps: Dict[str, Colormap] = {}
-
+class ColormapRegistry:
+    def __init__(self, *colormaps: Colormap):
+        self._categories: Dict[str, ColormapCategory] = {}
+        self._colormaps: Dict[str, Colormap] = {}
         # Add standard Matplotlib colormaps
-        for cm_name, cmap in matplotlib.colormaps.items():
-            template_category = _CMAP_NAME_TO_CAT.get(cm_name)
-            if template_category is None:
-                template_category = UNCATEGORIZED
-
-            category = categories.get(template_category.name)
-            if category is None:
-                category = ColormapCategory(template_category.name,
-                                            template_category.desc,
-                                            set())
-
-                categories[template_category.name] = category
-
-            category.cmap_names.add(cm_name)
-            cmaps[cm_name] = Colormap(category.name, cm_name, cmap)
-
+        self._register_mpl_cmaps()
         # Add Ocean colormaps, if any
-        if ocm is not None:
-            ocean_cat_template = OCEAN
-            ocean_cat_desc = _CAT_NAME_TO_DESC[ocean_cat_name]
-            ocean_cm_names = [k for k in ocm.__dict__.keys()
-                              if isinstance(getattr(ocm, k),
-                                            matplotlib.colors.Colormap)
-                              and not k.endswith("_r")]
-            categories[ocean_cat_name] = ocean_cat_desc, ocean_cm_names
-
+        self._register_ocm_cmaps()
         # Add custom colormaps, if any
-        if custom_cmaps:
-            custom_cat_name = CUSTOM[0]
-            custom_cat_desc = _CAT_NAME_TO_DESC[ocean_cat_name]
-            custom_cmap_names = []
-            for custom_cmap in custom_cmaps:
-                if not os.path.isfile(custom_cmap):
-                    LOG.error(f"Missing custom colormap file:"
-                              f" {custom_cmaps}")
-                    continue
-                try:
-                    cmap = _load_snap_colormap(custom_cmap)
-                except ValueError as e:
-                    LOG.warning(f'Detected invalid custom colormap'
-                                f' {custom_cmap!r}: {e}',
-                                exc_info=True)
-                custom_cmap_names.append(cmap)
-                cbar_list.append((cmap_name, _get_cmap_png_base64(cmap)))
+        for colormap in colormaps:
+            self.register_colormap(colormap)
 
-                cbar_list.append((cmap_name, _get_cmap_png_base64(cmap)))
-                cbar_list.append((new_name, _get_cmap_png_base64(new_cmap)))
+    @property
+    def categories(self) -> Dict[str, ColormapCategory]:
+        return self._categories
 
-            new_cmaps.append(
-                (cmap_category, cmap_description, tuple(cbar_list)))
+    @property
+    def colormaps(self) -> Dict[str, Colormap]:
+        return self._colormaps
 
-        return new_cmaps
+    def register_colormap(self, colormap: Colormap):
+        cat_name = colormap.cat_name or MISCELLANEOUS_CATEGORY.name
+        template_cat = TEMPLATE_CATEGORIES.get(cat_name)
+        if template_cat is None:
+            raise ValueError(f"Unknown colormap category {cat_name!r}")
+        category = self._get_category(template_cat)
+        category.cm_names.append(colormap.cm_name)
+        self._colormaps[colormap.cm_name] = colormap
+
+    def find_colormap(self, cm_name: str) -> Optional[Colormap]:
+        assert_instance(cm_name, str, name="cm_name")
+        # Strips suffixes from cm_name
+        cm_name, _, _ = parse_cm_name(cm_name)
+        return self._colormaps.get(cm_name)
+
+    def get_cmap(self,
+                 cm_name: str,
+                 default_cm_name: str = DEFAULT_CMAP_NAME,
+                 num_colors: Optional[int] = None) \
+            -> Tuple[str, matplotlib.colors.Colormap]:
+        assert_instance(cm_name, str, name="cm_name")
+        assert_instance(default_cm_name, str, name="default_cm_name")
+        if num_colors is not None:
+            assert_instance(default_cm_name, int, name="num_colors")
+
+        cm_name, reverse, alpha = parse_cm_name(cm_name)
+
+        colormap = self._colormaps.get(cm_name)
+        if colormap is None:
+            assert_in(default_cm_name, self._colormaps,
+                      name="default_cm_name")
+            cm_name = default_cm_name
+            colormap = self._colormaps[cm_name]
+
+        if reverse and alpha:
+            cmap = colormap.cmap_reversed_alpha
+        elif reverse:
+            cmap = colormap.cmap_reversed
+        elif alpha:
+            cmap = colormap.cmap_alpha
+        else:
+            cmap = colormap.cmap
+        if num_colors is not None:
+            try:
+                # noinspection PyProtectedMember
+                cmap = cmap._resample(num_colors)
+            except (ValueError, AttributeError, NotImplementedError):
+                pass
+        return cm_name, cmap
+
+    def to_json(self) -> List:
+        result = []
+        # Loop through TEMPLATE_CATEGORIES to preserve category order
+        for cat_name in TEMPLATE_CATEGORIES.keys():
+            category = self._categories.get(cat_name)
+            if category is not None:
+                result.append([
+                    category.name,
+                    category.desc,
+                    [
+                        [cm_name, self._colormaps[cm_name].cmap_png_base64]
+                        for cm_name in category.cm_names
+                    ]
+                ])
+        return result
+
+    def _get_category(self, template_cat: ColormapCategory):
+        category = self._categories.get(template_cat.name)
+        if category is None:
+            category = ColormapCategory(template_cat.name,
+                                        template_cat.desc,
+                                        cm_names=[])
+            self._categories[template_cat.name] = category
+        return category
+
+    def _register_mpl_cmaps(self):
+        for cm_name, cmap in matplotlib.colormaps.items():
+            if cm_name.startswith(OCEAN_PREFIX):
+                # We have a separate category for OCEAN maps
+                continue
+            if cm_name.endswith(REVERSED_SUFFIX):
+                # We register reversed maps, if any, separately
+                continue
+            template_cat = _CM_NAME_TO_TEMPLATE_CATEGORY.get(cm_name)
+            if template_cat is None:
+                template_cat = MISCELLANEOUS_CATEGORY
+            self.register_colormap(Colormap(
+                cm_name,
+                cat_name=template_cat.name,
+                cmap=cmap,
+                cmap_reversed=matplotlib.colormaps.get(
+                    cm_name + REVERSED_SUFFIX
+                )
+            ))
+
+    def _register_ocm_cmaps(self):
+        if cmo is None:
+            return
+        for cm_name in cmo.__dict__.keys():
+            if not isinstance(getattr(cmo, cm_name),
+                              matplotlib.colors.Colormap) \
+                    or cm_name.endswith(REVERSED_SUFFIX):
+                continue
+            self.register_colormap(Colormap(
+                cm_name,
+                cat_name=OCEAN_CATEGORY.name,
+                cmap=getattr(cmo, cm_name),
+                cmap_reversed=getattr(cmo,
+                                      cm_name + REVERSED_SUFFIX,
+                                      None)
+            ))
 
 
-def _get_alpha_cmap(cmap: matplotlib.colors.Colormap) \
-        -> Optional[matplotlib.colors.Colormap]:
+def parse_cm_name(cm_name) -> Tuple[str, bool, bool]:
+    alpha = cm_name.endswith(ALPHA_SUFFIX)
+    if alpha:
+        cm_name = cm_name[0:-len(ALPHA_SUFFIX)]
+    reverse = cm_name.endswith(REVERSED_SUFFIX)
+    if reverse:
+        cm_name = cm_name[0:-len(REVERSED_SUFFIX)]
+    return cm_name, reverse, alpha
+
+
+def get_reverse_cmap(cm_name: str, cmap: matplotlib.colors.Colormap) \
+        -> Tuple[str, matplotlib.colors.Colormap]:
+    new_cm_name = cm_name + REVERSED_SUFFIX
+    return new_cm_name, cmap.reversed(name=new_cm_name)
+
+
+def get_alpha_cmap(cm_name: str, cmap: matplotlib.colors.Colormap) \
+        -> Tuple[str, matplotlib.colors.Colormap]:
     """Add extra colormaps with alpha gradient.
     See http://matplotlib.org/api/colors_api.html
     """
-    cmap_name = cmap.name + '_alpha' \
-        if hasattr(cmap, 'name') else 'unknown'
+    new_cm_name = cm_name + ALPHA_SUFFIX
 
-    if isinstance(cmap, matplotlib.colors.LinearSegmentedColormap):
-        new_segmentdata = dict(cmap._segmentdata)
+    # noinspection SpellCheckingInspection
+    if isinstance(cmap, matplotlib.colors.LinearSegmentedColormap) \
+            and hasattr(cmap, "_segmentdata"):
+        # noinspection PyProtectedMember
+        new_segment_data = dict(cmap._segmentdata)
         # let alpha increase from 0.0 to 0.5
-        new_segmentdata['alpha'] = ((0.0, 0.0, 0.0),
-                                    (0.5, 1.0, 1.0),
-                                    (1.0, 1.0, 1.0))
-        return matplotlib.colors.LinearSegmentedColormap(
-            cmap_name, new_segmentdata
+        new_segment_data['alpha'] = ((0.0, 0.0, 0.0),
+                                     (0.5, 1.0, 1.0),
+                                     (1.0, 1.0, 1.0))
+        return new_cm_name, matplotlib.colors.LinearSegmentedColormap(
+            new_cm_name, new_segment_data
         )
 
     if isinstance(cmap, matplotlib.colors.ListedColormap):
@@ -285,6 +423,7 @@ def _get_alpha_cmap(cmap: matplotlib.colors.Colormap) \
         a_slope = 2.0 / cmap.N
         a = 0
         for i in range(len(new_colors)):
+            # noinspection PyUnresolvedReferences
             new_color = new_colors[i]
             if not isinstance(new_color, str):
                 if len(new_color) == 3:
@@ -296,31 +435,32 @@ def _get_alpha_cmap(cmap: matplotlib.colors.Colormap) \
             a += a_slope
             if a > 1.0:
                 a = 1.0
-        return matplotlib.colors.ListedColormap(
-            new_colors, name=cmap_name
+        return new_cm_name, matplotlib.colors.ListedColormap(
+            new_colors, name=new_cm_name
         )
 
     LOG.warning(
         f'Could not create alpha colormap for cmap of type {type(cmap)}'
     )
 
-    return None
+    return cm_name, cmap
 
 
-def _get_cmap_png_base64(cmap: matplotlib.colors.Colormap) -> str:
+def get_cmap_png_base64(cmap: matplotlib.colors.Colormap) -> str:
     gradient = np.linspace(0, 1, 256)
     gradient = np.vstack((gradient, gradient))
     image_data = cmap(gradient, bytes=True)
     image = Image.fromarray(image_data, 'RGBA')
 
-    # ostream = io.FileIO('../cmaps/' + cmap_name + '.png', 'wb')
-    # image.save(ostream, format='PNG')
-    # ostream.close()
+    # For testing, do:
+    # stream = io.FileIO('../cmaps/' + cmap_name + '.png', 'wb')
+    # image.save(stream, format='PNG')
+    # stream.close()
 
-    ostream = io.BytesIO()
-    image.save(ostream, format='PNG')
-    cbar_png_bytes = ostream.getvalue()
-    ostream.close()
+    stream = io.BytesIO()
+    image.save(stream, format='PNG')
+    cbar_png_bytes = stream.getvalue()
+    stream.close()
 
     cbar_png_data = base64.b64encode(cbar_png_bytes)
     cbar_png_bytes = cbar_png_data.decode('unicode_escape')
@@ -328,20 +468,24 @@ def _get_cmap_png_base64(cmap: matplotlib.colors.Colormap) -> str:
     return cbar_png_bytes
 
 
-def _load_snap_colormap(cpd_file_path: str):
-    points, log_scaled = _parse_snap_cpd_file(cpd_file_path)
+def load_snap_cpd_colormap(snap_cpd_path: str) -> Colormap:
+    points, log_scaled = _parse_snap_cpd_file(snap_cpd_path)
     samples = [sample for sample, _ in points]
-    # Uncomment, as soon as we know how to deal with this:
-    # if log_scaled:
-    #     norm = matplotlib.colors.LogNorm(min(samples), max(samples))
-    # else:
-    #     norm = matplotlib.colors.Normalize(min(samples), max(samples))
-    norm = matplotlib.colors.Normalize(min(samples), max(samples))
+    vmin, vmax = min(samples), max(samples)
+    if log_scaled:
+        norm = matplotlib.colors.LogNorm(vmin, vmax)
+    else:
+        norm = matplotlib.colors.Normalize(vmin, vmax)
     colors = list(zip(map(norm, samples),
                       ['#%02x%02x%02x' % color for _, color in points]))
-    cpd_name, _ = os.path.splitext(os.path.basename(cpd_file_path))
-    return matplotlib.colors.LinearSegmentedColormap.from_list(
-        cpd_name, colors
+    cm_name, _ = os.path.splitext(os.path.basename(snap_cpd_path))
+    return Colormap(
+        cm_name,
+        cat_name=CUSTOM_CATEGORY.name,
+        cmap=matplotlib.colors.LinearSegmentedColormap.from_list(
+            cm_name, colors
+        ),
+        norm=norm
     )
 
 
@@ -433,9 +577,75 @@ def get_norm(colortext):
     return norm, values
 
 
-def _check_if_exists(SNAP_CPD_LIST):
-    valid_path = []
-    for item in SNAP_CPD_LIST:
-        if os.path.isfile(item):
-            valid_path.append(item)
-    return tuple(valid_path)
+############################################################################
+# Deprecated API
+
+_REGISTRY = None
+_REGISTRY_JSON = None
+
+
+@deprecated(reason="Colormaps in xcube are no longer managed globally. Use"
+                   " xcube.util.cmaps.ColormapRegistry.to_json() instead.",
+            version="0.13.0")
+def get_cmaps() -> List:
+    """
+    Return a JSON-serializable tuple containing records of the form:
+     (<cmap-category>, <cmap-category-description>, <cmap-tuples>),
+    where <cmap-tuples> is a tuple containing records of the form
+    (<cmap-name>, <cbar-png-bytes>), and where
+    <cbar-png-bytes> are encoded PNG images of size 256 x 2 pixels,
+    :return: all known matplotlib color maps
+    """
+    registry = _get_registry()
+    global _REGISTRY_JSON
+    if _REGISTRY_JSON is None:
+        _REGISTRY_JSON = registry.to_json()
+    return _REGISTRY_JSON
+
+
+@deprecated(reason="Colormaps in xcube are no longer managed globally. Use"
+                   " xcube.util.cmaps.ColormapRegistry.get_cmap() instead.",
+            version="0.13.0")
+def get_cmap(cmap_name: str,
+             default_cmap_name: str = "viridis",
+             num_colors: Optional[int] = None) \
+        -> Tuple[str, matplotlib.colors.Colormap]:
+    """
+    Get color mapping for color bar name *cmap_name*.
+
+    If *num_colors* is a positive integer,
+    a resampled mapping will be returned that contains *num_colors*
+    color entries.
+
+    If *cmap_name* is not defined, *default_cmap_name* is used.
+    If *default_cmap_name* is undefined too, a ValueError is raised.
+
+    Otherwise, a tuple (actual_cmap_name, cmap) is returned.
+
+    :param cmap_name: Color bar name.
+    :param num_colors: Number of colours in returned color mapping.
+    :param default_cmap_name: Default color bar name.
+    :return: A tuple (actual_cmap_name, cmap).
+    """
+    registry = _get_registry()
+    return registry.get_cmap(cmap_name,
+                             default_cm_name=default_cmap_name,
+                             num_colors=num_colors)
+
+
+@deprecated(reason="Colormaps in xcube are no longer managed globally."
+                   " This function is obsolete.",
+            version="0.13.0")
+def ensure_cmaps_loaded():
+    """
+    Loads all color maps from matplotlib and registers additional ones,
+    if not done before.
+    """
+    _get_registry()
+
+
+def _get_registry() -> ColormapRegistry:
+    global _REGISTRY
+    if _REGISTRY is None:
+        _REGISTRY = ColormapRegistry()
+    return _REGISTRY